[MNT] move the shared grid search engine to sktime.base._panel
