+++ sktime/base/_panel/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for shared panel estimator base classes."""
